Reworked BuildWrapper to SimpleBuildWrapper and implemented JenkinsRule based test for Workflow support.
