+++ pom.xml
@@ -81,8 +81,9 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>4.8.1</version>
+      <version>4.11</version>
       <scope>test</scope>
+      <type>jar</type>
     </dependency>
     <dependency>
       <groupId>org.fusesource.jansi</groupId>
@@ -93,6 +94,7 @@
       <groupId>org.jenkins-ci.plugins.workflow</groupId>
       <artifactId>workflow-aggregator</artifactId>
       <version>1.8</version>
+      <scope>test</scope>
     </dependency>
   </dependencies>
