+++ src/test/java/hudson/plugins/ansicolor/AnsiColorBuildWrapperTest.java
@@ -1,11 +1,22 @@
 package hudson.plugins.ansicolor;
 
+import hudson.Functions;
+import java.io.StringWriter;
 import static org.hamcrest.CoreMatchers.instanceOf;
+import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.job.WorkflowJob;
+import org.jenkinsci.plugins.workflow.job.WorkflowRun;
 import org.junit.Test;
 import static org.junit.Assert.*;
+import org.junit.Assume;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.runners.model.Statement;
+import org.jvnet.hudson.test.BuildWatcher;
+import org.jvnet.hudson.test.RestartableJenkinsRule;
 
 public class AnsiColorBuildWrapperTest {
-    
+
     public AnsiColorBuildWrapperTest() {
     }
 
@@ -14,7 +25,7 @@ public class AnsiColorBuildWrapperTest {
         AnsiColorBuildWrapper instance = new AnsiColorBuildWrapper(null, 1, 2);
         assertEquals("xterm", instance.getColorMapName());
     }
-    
+
     @Test
     public void testGetColorMapNameVga() {
         AnsiColorBuildWrapper instance = new AnsiColorBuildWrapper("vga", 1, 2);
@@ -26,5 +37,34 @@ public class AnsiColorBuildWrapperTest {
         AnsiColorBuildWrapper ansiColorBuildWrapper = new AnsiColorBuildWrapper(null, 1, 2);
         assertThat(ansiColorBuildWrapper, instanceOf(AnsiColorBuildWrapper.class));
     }
-    
+
+    @ClassRule
+    public static BuildWatcher buildWatcher = new BuildWatcher();
+    @Rule
+    public RestartableJenkinsRule story = new RestartableJenkinsRule();
+
+    @Test
+    public void testWorkflowWrap() throws Exception {
+        story.addStep(new Statement() {
+
+            @Override
+            public void evaluate() throws Throwable {
+                Assume.assumeTrue(!Functions.isWindows());
+                WorkflowJob p = story.j.jenkins.createProject(WorkflowJob.class, "p");
+                p.setDefinition(new CpsFlowDefinition(
+                        "node {\n"
+                        + "  wrap([$class: 'AnsiColorBuildWrapper', 'colorMapName': 'XTerm', 'defaultFg': 1, 'defaultBg': 2]) {\n"
+                        + "    sh(\"\"\"#!/bin/bash\n"
+                        + "      echo -e '\\\\e[31mred\\\\e[0m'\"\"\"\n"
+                        + "    )\n"
+                        + "  }\n"
+                        + "}"
+                ));
+                story.j.assertBuildStatusSuccess(p.scheduleBuild2(0));
+                StringWriter writer = new StringWriter();
+                p.getLastBuild().getLogText().writeHtmlTo(0L, writer);
+                assertTrue(writer.toString().matches("(?s).*<span style=\"color: #CD0000;\">red</span>.*"));                        
+            }
+        });
+    }
 }
